refactor(moodpage): organize files
